+++ modules/non-nixos.nix
@@ -1,8 +1,8 @@
-{ pkgs, ... }:
-
 {
-  home.packages = builtins.attrValues {
-     # Necessary for non-NixOS systems which won't have the flakiest version of Nix
-    inherit (pkgs) nix;
+  hmModule = { pkgs, ... }: {
+    home.packages = builtins.attrValues {
+       # Necessary for non-NixOS systems which won't have the flakiest version of Nix
+      inherit (pkgs) nix;
+    };
   };
 }