+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- Changed `non-nixos` module into a regular module
+- Added `graphical` module to be the base of any graphical system
+- Added `imports` attribute for modules to import other modules recursively
 - Added `gsc` alias
 - Added `spotify`
 - Added `grias` alias